+++ attributes/default.rb
@@ -20,3 +20,4 @@
 
 default['chef_dk']['version'] = 'latest'
 default['chef_dk']['package_url'] = nil
+default['chef_dk']['global_shell_init'] = false