+++ spec/libraries/provider_chef_dk_mac_os_x_spec.rb
@@ -97,4 +97,10 @@ describe Chef::Provider::ChefDk::MacOsX do
       expect(provider.send(:package_file_extension)).to eq('.dmg')
     end
   end
+
+  describe '#bashrc_file' do
+    it 'returns "bashrc"' do
+      expect(provider.send(:bashrc_file)).to eq('/etc/bashrc')
+    end
+  end
 end