+++ spec/libraries/provider_chef_dk_windows_spec.rb
@@ -88,4 +88,11 @@ describe Chef::Provider::ChefDk::Windows do
       expect(provider.send(:package_file_extension)).to eq('.msi')
     end
   end
+
+  describe '#bashrc_file' do
+    it 'raises an exception' do
+      expected = Chef::Exceptions::UnsupportedPlatform
+      expect { provider.send(:bashrc_file) }.to raise_error(expected)
+    end
+  end
 end