+++ test/integration/default/serverspec/localhost/package_spec.rb
@@ -19,17 +19,20 @@
 
 require 'spec_helper'
 
-case os[:family]
-when 'Darwin'
-  describe package('com.getchef.pkg.chefdk') do
-    it 'is installed' do
-      expect(described_class).to be_installed.by(:pkgutil).with_version('0.2.0')
+describe 'Chef-DK package' do
+  case os[:family]
+  when 'Darwin'
+    describe package('com.getchef.pkg.chefdk') do
+      it 'is installed' do
+        expect(described_class).to be_installed.by(:pkgutil)
+          .with_version('0.2.0')
+      end
     end
-  end
-else
-  describe package('chefdk') do
-    it 'is installed' do
-      expect(described_class).to be_installed.with_version('0.2.0-2')
+  else
+    describe package('chefdk') do
+      it 'is installed' do
+        expect(described_class).to be_installed.with_version('0.2.0-2')
+      end
     end
   end
 end