+++ test/integration/global_shell_init/serverspec/localhost/package_spec.rb
@@ -1,0 +1,1 @@
+../../../default/serverspec/localhost/package_spec.rb