+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Chef-DK Cookbook CHANGELOG
 ==========================
 
+v?.?.? (????-??-??)
+-------------------
+
 v1.0.2 (2014-07-31)
 -------------------
 * Fix bug with OS X package reinstalling on every Chef run